# docs/openapi.json
{
  "openapi": "3.0.3",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created."
          }
        },
        "summary": "Register a user with the USER role",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Authenticate with email and password",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login/mfa": {
      "post": {
        "operationId": "AuthController_loginWithMfa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Complete an MFA login challenge",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "refreshCookie": [],
            "csrfToken": []
          }
        ],
        "summary": "Rotate the refresh session and issue an access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "refreshCookie": [],
            "csrfToken": []
          }
        ],
        "summary": "Revoke the current refresh session",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Return the authenticated identity",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/tickets/{ticketId}/attachments": {
      "post": {
        "operationId": "AttachmentsController_upload",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Attach a validated JPEG or PNG image to a visible ticket",
        "tags": [
          "attachments"
        ]
      },
      "get": {
        "operationId": "AttachmentsController_list",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 10000,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 50,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List metadata for attachments on a visible ticket",
        "tags": [
          "attachments"
        ]
      }
    },
    "/attachments/{id}/content": {
      "get": {
        "operationId": "AttachmentsController_download",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download an authorized attachment as an opaque file",
        "tags": [
          "attachments"
        ]
      }
    },
...
    "/auth/mfa/status": {
      "get": {
        "operationId": "MfaController_status",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Return MFA status for the authenticated user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/mfa/setup": {
      "post": {
        "operationId": "MfaController_setup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaSetupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Secret and authenticator URI. Display the secret only during setup."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate a TOTP secret for optional MFA setup",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/mfa/verify-setup": {
      "post": {
        "operationId": "MfaController_verifySetup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyMfaSetupDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify the TOTP code and enable MFA",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/mfa/disable": {
      "post": {
        "operationId": "MfaController_disable",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableMfaDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Disable MFA after verifying the current TOTP code",
        "tags": [
          "auth"
        ]
      }
    }
  },
  "info": {
    "title": "Secure Service Desk API",
    "description": "Secure, modular service desk API for portfolio demonstration. The MVP is single-tenant and manages identity, tickets, files, audit, notifications and reports in code.",
    "version": "0.1.0",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/license/mit/"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local demo reverse proxy"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Standalone API process"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "service_desk_refresh"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token",
        "description": "Double-submit CSRF token returned by register or login."
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128,
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128,
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "MfaLoginDto": {
        "type": "object",
        "properties": {
          "challengeToken": {
            "type": "string",
            "description": "Opaque one-time challenge returned by /auth/login."
          },
          "code": {
            "type": "string",
            "description": "Six-digit TOTP code.",
            "example": "123456"
          }
        },
        "required": [
          "challengeToken",
          "code"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "MfaSetupDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "Current account password.",
            "writeOnly": true
          }
        },
        "required": [
          "password"
        ]
      },
      "VerifyMfaSetupDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "Current account password.",
            "writeOnly": true
          },
          "code": {
            "type": "string",
            "description": "Current six-digit TOTP code.",
            "example": "123456"
          }
        },
        "required": [
          "password",
          "code"
        ]
      },
      "DisableMfaDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "Current account password."
          },
          "code": {
            "type": "string",
            "description": "Current six-digit TOTP code.",
            "example": "123456"
          }
        },
        "required": [
          "password",
          "code"
        ]
      }
    }
  }
}